requirements.txt
asgiref==3.4.1
attrs==21.2.0
autopep8==1.4.4
awsebcli==3.20.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
blessed==1.18.1
botocore==1.20.101
bs4==0.0.1
cached-property==1.5.2
cement==2.8.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.3
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-allauth==0.44.0
django-appconf==1.0.4
django-compressor==2.4.1
django-dotenv==1.4.2
django-sass-processor==1.0.1
djangorestframework==3.12.4
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
idna==2.10
importlib-metadata==4.5.0
jmespath==0.10.0
jsonschema==3.2.0
libsass==0.21.0
mysqlclient==2.0.1
oauthlib==3.1.1
paramiko==2.7.2
pathspec==0.5.9
pycodestyle==2.5.0
pycparser==2.20
PyJWT==2.1.0
PyMySQL==1.0.2
PyNaCl==1.4.0
PyPDF2==1.26.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
rcssmin==1.0.6
requests==2.25.1
requests-oauthlib==1.3.0
rjsmin==1.1.0
selenium==3.141.0
semantic-version==2.8.5
six==1.14.0
soupsieve==2.2.1
sqlparse==0.4.1
termcolor==1.1.0
texttable==1.6.3
typing-extensions==3.10.0.0
urllib3==1.25.11
wcwidth==0.1.9
websocket-client==0.59.0
zipp==3.4.1
